+++ src/test/java/com/picimako/mockitools/completion/MockitoolsCodeCompletionTestBase.java
@@ -1,0 +1,30 @@
+//Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+package com.picimako.mockitools.completion;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.intellij.codeInsight.completion.CompletionType;
+import com.picimako.mockitools.MockitoolsTestBase;
+
+/**
+ * Base class for testing code completion.
+ */
+public abstract class MockitoolsCodeCompletionTestBase extends MockitoolsTestBase {
+
+    /**
+     * Tests basic code completion in the provided file text at the marked caret position,
+     * and validates if the completion contains exactly the expected items.
+     *
+     * @param fileName        the file name to configure the test data in
+     * @param text            the file text on which the code completion is invoked
+     * @param completionItems the list of items that are expected to be in the displayed completion list
+     */
+    protected void doTestCodeCompletionContains(String fileName, String text, String... completionItems) {
+        assertThat(text).withFailMessage("File text must contain <caret> to invoke code completion.").contains("<caret>");
+
+        getFixture().configureByText(fileName, text);
+        getFixture().complete(CompletionType.BASIC);
+
+        assertThat(getFixture().getLookupElementStrings()).containsExactlyInAnyOrder(completionItems);
+    }
+}